Fix shutdown bug
The issue here was that VertexData tried to deallocate the memory after the GLFW/OpenGL instance (not totally sure which) had already been destroyed.

The fix is to rearrange the struct members so that the window is deallocated after the other things that implicitly depend on it, plus a small fix to stop things from failing to write to the event queue after the window has been closed. (This bug already existed, it was just masked by the deallocation issue.)